changed amounts to be represented in cents

+++ personalfinance/models.py
@@ -8,7 +8,7 @@ class Transaction(models.Model):
     name = models.CharField(max_length=100)
     category = models.CharField(max_length=50)
     date = models.DateTimeField('Transaction Date')
-    amount = models.FloatField()
+    amount = models.IntegerField()
     recurring = models.BooleanField()
     user = models.ForeignKey(User, on_delete=models.CASCADE, related_name='Transactions')
 
@@ -30,8 +30,8 @@ class Budget(models.Model):
 # pot model
 class Pot(models.Model):
     name = models.CharField(max_length=50)
-    target = models.FloatField()
-    total = models.FloatField(default=0.00)
+    target = models.IntegerField()
+    total = models.IntegerField(default=0)
     user = models.ForeignKey(User, on_delete=models.CASCADE, related_name='Pots')
     theme = models.CharField(max_length=7)
